requirements.txt
streamlit==1.32.2
torch==2.2.1
torchvision==0.17.1
numpy
opencv-python-headless
pillow
matplotlib
streamlit-image-comparison
